+++ template/vitest.config.ts
@@ -1,5 +1,4 @@
 /// <reference types="vitest" />
-/// <reference types="vite/client" />
 
 import react from "@vitejs/plugin-react";
 import vue from "@vitejs/plugin-vue";
@@ -8,8 +7,10 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   plugins: [
     /** TODO: Add Angular testing */
-    react(),
-    vue(),
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    react() as any,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    vue() as any,
   ],
   test: {
     environment: "happy-dom",