+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added ESLint linting.
 - Added Prettier formatting.
 - Added Playwright end-to-end testing.
 - Added Testing Library integration testing for React and Vue.